Expand README documentation for ServiceModel usage
The documentation in README.md file has been significantly expanded to provide a more comprehensive guide on how to use the ServiceModel. Detailed instructions and code examples have been given to illustrate the usage of new features. Naming changes from 'CastMethod' to 'CastUsing' and 'CastToClasses' to 'ArrayOf' are also reflected in the updated README.