Feat/validation (#63)
* Implement strict mode and validation in ServiceModel

The ServiceModel.php file was updated to include strict mode, and a validation method was included for non-null properties. Allowing for strict mode facilitates rigorous checks on properties. The validation mechanism enables the application to throw an exception when mandatory fields are not correctly set, thereby improving overall data integrity.

* Add validation feature and rename MyClass to MyModel

This commit implements a new validation feature in the README file and renames several instances of 'MyClass' to 'MyModel'. The validation feature provides a mechanism for checking if required properties are initialized in the model, improving the robustness of the ServiceModel. The renaming of the model class enhances the readability and understanding of the code.

* Add new 'Strict' trait to ServiceModel

This commit introduces a new trait 'Strict' to the ServiceModel in the file src/Strict.php. This trait is added to enhance. Details about how it will be used or how it will work have yet to be implemented.

* Add ValidationException to ServiceModel exceptions

This commit introduces a new 'ValidationException' class under ServiceModel/Exceptions. This class extends the base Exception class. Details about its usage or how it will work are yet to be implemented.

* Add new models and tests for validation and strict checks

This commit introduces 'StrictChild', 'StrictClass', and 'ValidationClass' new models in the Zerotoprod/AppServiceModel. Validation tests have been added to check validation and strict behaviour, and constant properties have been defined for the Child model.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>