Add Instantiated Class Support (#67)
* Add condition to check object type before assignment

The update introduces a condition to check if a value is an instance of a model class before assignment. If the current value is an instance of the expected class, it gets assigned directly without creating a new instance. This improves performance by avoiding unnecessary object creation.

* Add CustomCastWithConstructor test model

This commit introduces a new test model, CustomCastWithConstructor, for testing custom constructor casting. This new class contains a constructor that takes a single string argument, thus expanding the testing parameters for model casting within the application's services.

* Update test descriptions in CastToClassTest

This commit renames a test description for clarity and adds a new test case related to the 'CustomCastWithConstructor' test model. This update aims to provide better context and diversify testing for class casting.

* Add new custom class with constructor in test model

A new custom class 'custom_class_with_constructor' that includes a constructor has been added to the CastToClass test model. This inclusion is done to expand test coverage and ensure the correct handling of classes with constructors during casting.

---------

Co-authored-by: david_smith <david_smith@sweetwater.com>